Quick teardown and creation of VMs can cause stale dns entires in VMs.
If user tear down the VMs and recreate the VMs again shortly,
newly created VMs might retain the stale dns cache entry
for hostname (mapping to previosly created VM's IP address).
This causes issue with the OVN kubernetes setup because it
relies on - $ getent ahostsv4 <hostname>- command
to get the host ip address and that returnes the
stale IP address for the hostname and that results
in cluster setup failure.

This Patch provides the workaround fix that will
reload the dnsmasq configuration once VMs are
created so VMs DNS cache gets the latest dns records.

Signed-off-by: Anil Vishnoi <vishnoianil@gmail.com>